requirements.txt
pyyaml==5.3.1
locust==0.14.6
